# api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Agent 平台 API
  version: 1.0.0
  contact:
    name: AI Agent 平台团队
    email: platform@example.com
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    AI Agent 平台的 HTTP API 规范，是前后端联调的**唯一事实来源**。

    关键约定（详见《架构设计文档》第六章）：
    - 除事件流接口外，所有响应使用统一信封 `{code, message, data, request_id}`
    - `code = 0` 表示成功；非 0 为五位业务错误码（AABBB：前两位模块，后三位序号）
    - 列表字段无数据时必须返回 `[]`，**禁止返回 null**
    - 时间一律 RFC3339 带时区；对外 ID 一律字符串（避免 JS Number 精度丢失）
    - 创建类接口支持 `Idempotency-Key` 请求头
servers:
  - url: https://api.example.com/api/v1
    description: 生产环境
  - url: http://localhost:8080/api/v1
    description: 本地开发
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: 注册
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, display_name]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8, writeOnly: true }
                display_name: { type: string, maxLength: 64 }
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/AuthResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409':
          description: 邮箱已注册
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
  /auth/login:
    post:
      tags: [Auth]
      summary: 登录
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, writeOnly: true }
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/AuthResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ────────────────────────── 应用广场 ──────────────────────────
  /marketplace/listings:
    get:
      tags: [Marketplace]
      summary: 广场浏览
      description: |
        返回**黑盒安全**的展示信息，绝不包含 persona、编排图等内部定义。
      operationId: browseMarketplace
      parameters:
        - name: resource_type
          in: query
          schema: { $ref: '#/components/schemas/ListingResourceType' }
        - name: q
          in: query
          description: 按名称/描述搜索
          schema: { type: string }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [items, has_more]
                        properties:
                          items:
                            type: array
                            items: { $ref: '#/components/schemas/ListingSummary' }
                          next_cursor: { type: [string, 'null'] }
                          has_more: { type: boolean }
        '401': { $ref: '#/components/responses/Unauthorized' }

    post:
      tags: [Marketplace]
      summary: 发布资源到广场
      description: |
        发布前做**递归依赖闭包校验**：被发布资源传递依赖的每一层
        （Bundle → Agent → MCP/Skill/Tool）都必须已发布，任一层存在
        未发布的私有资源即返回 422 + 错误码 70001。

        校验按精确版本匹配——Bundle 引用的是 Agent 的具体版本，
        必须是「该版本」已发布，不是「该 Agent 有任意版本发过」。

        存在多个未发布依赖时一次性全部返回，`details` 给出完整依赖路径。
        循环依赖返回 70008。
      operationId: publishListing
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resource_type, resource_ref, version, display_meta]
              properties:
                resource_type: { $ref: '#/components/schemas/ListingResourceType' }
                resource_ref:
                  type: string
                  description: 作者空间内的资源标识
                version: { type: string, examples: ['2.0'] }
                display_meta: { $ref: '#/components/schemas/ListingDisplayMeta' }
                changelog:
                  type: string
                  description: |
                    本版本的更新说明，由作者手写。**不提供自动 diff**——
                    diff 会泄露黑盒内容。
      responses:
        '201':
          description: 发布成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/ListingDetail' }
        '422':
          description: 依赖完整性校验失败（70001）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
              examples:
                privateDep:
                  value:
                    code: 70001
                    message: '该资源引用了未发布的私有资源，无法发布'
                    data: null
                    request_id: 'req-a1b2c3d4'
                    details:
                      - field: 'agents[1].capabilities.tools[2]'
                        reason: '依赖路径 web-app-builder → architect@1.0 → mcp/internal-search@1.0 未发布'
                      - field: 'agents[0]'
                        reason: '依赖路径 web-app-builder → product_manager@1.2 未发布'
  /marketplace/listings/{ref}:
    parameters:
      - name: ref
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Marketplace]
      summary: 广场资源详情
      description: |
        黑盒资源仅返回 `display_meta` 与 `constraints_summary`。
        **`constraints_summary` 是黑盒规则的必要例外**——不暴露 persona 和
        编排图，但必须暴露执行约束，否则订阅者无法预估 Token 成本。
      operationId: getListing
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/ListingDetail' }
        '404':
          description: listing 不存在或已下架（70002）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
  /marketplace/listings/{ref}/report:
    parameters:
      - name: ref
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [Operations]
      summary: 举报该资源
      description: 提交后进入运营中心的管理员举报队列（GET /moderation/reports）。
      operationId: reportListing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, examples: ['描述与实际功能严重不符'] }
      responses:
        '201':
          description: 已提交
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Report' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404':
          description: listing 不存在（70002）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
  /marketplace/listings/{id}/unpublish:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [Marketplace]
      summary: 停止分发
      description: |
        停止分发后广场不再展示，但**不影响存量订阅者**（他们继续可用）。

        若该资源正被其他**已发布**资源依赖，拒绝停止分发（70007）——
        否则上游资源的依赖闭包会断裂，订阅了上游的用户会突然跑不了。
        响应的 details 列出具体依赖方。
      operationId: unpublishListing
      responses:
        '204':
          description: 已停止分发
        '409':
          description: 被其他已发布资源依赖（70007）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
              examples:
                depended:
                  value:
                    code: 70007
                    message: '该资源正被其他已发布资源依赖，无法停止分发'
                    data: null
                    request_id: 'req-a1b2c3d4'
                    details:
                      - field: 'depended_by'
                        reason: 'Bundle web-app-builder@2.0 正在依赖'
  /marketplace/listings/{id}/subscribe:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [Marketplace]
      summary: 订阅
      description: |
        订阅关系绑定到**该 listing 的具体版本**（快照隔离）。
        作者后续发布新版本不影响本次订阅，订阅者需显式升级。
        订阅动作会把对应资源版本标记为不可变，防止作者事后修改。
      operationId: subscribeListing
      responses:
        '201':
          description: 订阅成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Subscription' }
        '404':
          description: listing 不存在或已停止分发（70002）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        '409':
          description: 已订阅该版本（70004）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
  /marketplace/subscriptions:
    get:
      tags: [Marketplace]
      summary: 我的订阅
      operationId: listSubscriptions
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [items, has_more]
                        properties:
                          items:
                            type: array
                            items: { $ref: '#/components/schemas/Subscription' }
                          next_cursor: { type: [string, 'null'] }
                          has_more: { type: boolean }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /marketplace/subscriptions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    delete:
      tags: [Marketplace]
      summary: 退订
      operationId: unsubscribe
      responses:
        '204':
          description: 退订成功（历史运行记录仍可查）
        '404': { $ref: '#/components/responses/NotFound' }
  /marketplace/subscriptions/{id}/upgrade:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [Marketplace]
      summary: 升级订阅到新版本
      description: 显式操作，不会自动发生。升级后旧版本的运行历史仍可查。
      operationId: upgradeSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_version]
              properties:
                target_version: { type: string, examples: ['2.0'] }
      responses:
        '200':
          description: 升级成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Subscription' }
        '404':
          description: 目标版本不存在（70002）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }

  # ────────────────────────── 资源中心 ──────────────────────────
  /resources:
    get:
      tags: [Resources]
      summary: 资源列表
      operationId: listResources
      parameters:
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/ResourceType'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [items, has_more]
                        properties:
                          items:
                            type: array
                            items: { $ref: '#/components/schemas/Resource' }
                          next_cursor: { type: [string, 'null'] }
                          has_more: { type: boolean }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      tags: [Resources]
      summary: 注册资源
      operationId: createResource
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, ref, config]
              properties:
                type: { $ref: '#/components/schemas/ResourceType' }
                ref:
                  type: string
                  pattern: '^[a-z][a-z0-9_-]*$'
                  description: Agent DSL 中引用该资源使用的 ID
                  examples: ['internal-search']
                display_name: { type: string }
                config:
                  type: object
                  description: 因资源类型而异的连接配置
                  additionalProperties: true
                  examples:
                    - { endpoint: 'https://mcp.internal/search', auth_type: 'bearer' }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Resource' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409':
          description: ref 重复（错误码 30003）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
  /resources/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    patch:
      tags: [Resources]
      summary: 更新资源（含启用/禁用）
      operationId: updateResource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name: { type: string }
                config: { type: object, additionalProperties: true }
                status: { $ref: '#/components/schemas/Status' }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Resource' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ────────────────────────── Agent ──────────────────────────
  /resources/{id}/delete-check:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Resources]
      summary: 删除前占用检查
      description: 返回该资源被哪些 Agent 引用，供前端在确认弹窗中展示影响范围
      operationId: checkResourceDeletable
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          deletable: { type: boolean }
                          referenced_by:
                            type: array
                            items:
                              type: object
                              properties:
                                type: { type: string }
                                ref: { type: string }
                                version: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /resources/{id}/kb/documents:
    parameters:
      - name: id
        in: path
        required: true
        description: 一个 knowledge_base 类型资源的 id；其他类型返回 404
        schema: { type: string }
    get:
      tags: [Resources]
      summary: 已摄入的文档列表（按 source_ref 聚合，不是按 chunk）
      operationId: listKnowledgeBaseDocuments
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/KnowledgeBaseDocument' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Resources]
      summary: 摄入一份文档（分块、向量化、存储）
      description: 用相同 source_ref 重新摄入即为更新——旧的 chunk 会先被整体替换
      operationId: ingestKnowledgeBaseDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source_ref, content]
              properties:
                source_ref: { type: string, description: 文档标识，如文件名或标题 }
                content: { type: string }
      responses:
        '201':
          description: 摄入成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          chunk_count: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422':
          description: embedding 调用失败或维度不匹配
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
  /resources/{id}/kb/documents/{source_ref}:
    delete:
      tags: [Resources]
      summary: 删除一份文档的全部 chunk
      operationId: deleteKnowledgeBaseDocument
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: source_ref
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: 已删除 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /resources/{id}/kb/search:
    post:
      tags: [Resources]
      summary: 检索预览——手动验证摄入的内容能不能被合理召回
      operationId: searchKnowledgeBase
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query: { type: string }
                top_k: { type: integer, default: 5 }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/KnowledgeBaseSearchResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /agents:
    get:
      tags: [Agents]
      summary: Agent 列表（每个 ref 返回最新启用版本）
      operationId: listAgents
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [items, has_more]
                        properties:
                          items:
                            type: array
                            items: { $ref: '#/components/schemas/Agent' }
                          next_cursor: { type: [string, 'null'] }
                          has_more: { type: boolean }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      tags: [Agents]
      summary: 创建 Agent（新建或新增版本）
      description: |
        请求体中的 `definition` 会用 Agent DSL 的 JSON Schema 严格校验。
        校验失败返回 400 + 错误码 40001，`details` 中给出具体字段路径。
      operationId: createAgent
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [definition]
              properties:
                definition: { $ref: '#/components/schemas/AgentDefinition' }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Agent' }
        '400':
          description: |
            DSL 校验失败（40001）或引用了被禁用的资源（30002）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
              examples:
                dslInvalid:
                  summary: DSL 字段不合法
                  value:
                    code: 40001
                    message: 'Agent 定义不符合 Schema'
                    data: null
                    request_id: 'req-a1b2c3d4'
                    details:
                      - field: 'model.provider'
                        reason: '必须是 anthropic / openai / google / custom 之一'
  /agents/{ref}:
    parameters:
      - name: ref
        in: path
        required: true
        schema: { type: string }
    delete:
      tags: [Agents]
      summary: 删除 Agent
      description: |
        删除前校验两类占用：
        1. 该 Agent 的任一版本已被发布且有存量订阅者 → 拒绝（快照隔离要求），只能停止分发
        2. 被其他 Bundle 引用中 → 拒绝，并在 details 列出具体 Bundle
        删除不影响历史运行记录的可查性。
      operationId: deleteAgent
      responses:
        '204':
          description: 删除成功
        '409':
          description: 存在占用无法删除（70005 已被订阅 / 40004 被引用）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
              examples:
                inUse:
                  value:
                    code: 40004
                    message: '该 Agent 正被其他 Bundle 引用，无法删除'
                    data: null
                    request_id: 'req-a1b2c3d4'
                    details:
                      - field: 'referenced_by'
                        reason: 'Bundle web-app-builder v2.0 正在引用'
  /agents/{ref}/versions:
    get:
      tags: [Agents]
      summary: 某个 Agent 的全部版本
      operationId: listAgentVersions
      parameters:
        - name: ref
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Agent' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ────────────────────────── Bundle ──────────────────────────
...
  /bundles/{ref}:
    parameters:
      - name: ref
        in: path
        required: true
        schema: { type: string }
    delete:
      tags: [Bundles]
      summary: 删除 Bundle
      description: 已被订阅的版本不可删除，只能停止分发。删除不影响历史运行记录。
      operationId: deleteBundle
      responses:
        '204':
          description: 删除成功
        '409':
          description: 已被订阅无法删除（70005）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }

  # ────────────────────────── Run ──────────────────────────
...
  /runs/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
        examples: { run: { value: 'run-9e67931d5c38024e' } }
    get:
      tags: [Runs]
      summary: 运行详情（含最新 shared_state）
      operationId: getRun
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/RunDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /runs/{id}/cancel:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [Runs]
      summary: 停止运行
      description: |
        用户主动终止一次运行，避免继续消耗 Token。停止在 5 秒内生效，
        不等待当前 Agent 执行完。已产生的产出与事件完整保留，
        **已消耗的 Token 照常计费**（无法退回）。
      operationId: cancelRun
      responses:
        '204':
          description: 已发出停止指令
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: 运行已结束（50002）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Web 控制台用户会话，有效期 2h
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 服务端调用 / CI 集成，格式 `ApiKey <key>`
...
  responses:
    BadRequest:
      description: 请求参数格式错误或 Schema 校验失败
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
...
  schemas:
...
    Envelope:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: integer
          description: 0 为成功；非 0 为五位业务错误码
          examples: [0]
        message: { type: string, examples: ['ok'] }
        data:
          description: 业务数据，无返回内容时为 null
        request_id:
          type: string
          description: 全链路追踪 ID，等同 OTel trace_id
          examples: ['req-a1b2c3d4']
        details:
          type: array
          description: 字段级错误，表单场景使用
          items:
            type: object
            required: [field, reason]
            properties:
              field: { type: string }
              reason: { type: string }
...
    AuthResult:
      type: object
      required: [access_token, refresh_token, user]
      properties:
        access_token: { type: string, description: 'JWT，有效期 2h' }
        refresh_token: { type: string, description: '有效期 7d' }
        user: { $ref: '#/components/schemas/User' }
...